新增权限管理，按注入的Token来限制用户登入后的权限，仅有管理员才可开放数据库操作权限。

+++ manage/src/main/java/com/example/demo/control/PersonControl.java
@@ -29,17 +29,16 @@ public class PersonControl {
 
     public static Map<Object,Object> map = new HashMap<>();
 
+
     /*
     用户登入(账号密码)
      */
-    @PostMapping(value = "/login")
-    public CreditsEntity userlogin(@Validated @RequestBody UserLogin userLogin) {
+    @PostMapping(value = "/login", produces = "application/json")
+    public CreditsEntity userLogin(@Validated @RequestBody UserLogin userLogin) {
         UserLogin login = loginRepository.findByUsername(userLogin.getUsername());
         if (login.getPassword().equals(userLogin.getPassword())) {
             System.out.println("登陆成功");
-            TimeSet timeSet = new TimeSet();
-            timeSet.timeStart(login.getId());
-
+            //Token数据库操作方法loginToken()
             return loginToken(login);
         } else {
             System.out.println("登陆失败");
@@ -48,7 +47,7 @@ public class PersonControl {
     }
 
     //获取已登入Token用户信息
-    @PostMapping(value = "/token")
+    @PostMapping(value = "/token", produces = "application/json")
     public Object loginByToken(@Validated @RequestBody CreditsEntity requestObj){
         if(map.get("token")==null){
             map.put("token","");
@@ -57,18 +56,28 @@ public class PersonControl {
         if(map.get("token").equals(requestObj.getToken()) && requestObj.getToken()!=null){
             return map.get("user");
         }
-        return "Token无效";
+        Token token = tokenRepository.findByToken(requestObj.getToken());
+        if(token == null){
+            return "Token无效";
+        }
+        Person person = personRepository.findOne(token.getId());
+        if(person == null){
+            return "无该人员信息";
+        }
+        return person;
     }
 
     /*
     查询所有人员信息
      */
     @GetMapping(value = "/persons")
-    public List<CreditsEntity> person_list() {
+    public Object personList(@RequestParam("access_token") String token) {
+        if(verifyToken(token)==false){
+            return "无请求权限，操作错误";
+        }
         //返回类数组初始化
         List<CreditsEntity> creditsEntities = new ArrayList<>();
         List<Person> persons = personRepository.findAll();
-
         for (Person person : persons) {
             //返回类cred填充数据
             CreditsEntity cred = new CreditsEntity();
@@ -79,6 +88,10 @@ public class PersonControl {
             for (Person_Position p : person_positions) {
                 Position position = positionRepository.myFindPo(p.getPositionId());
                 cred.getPosition().add(position);
+            }
+            Token mytoken = tokenRepository.findOne(person.getId());
+            if(mytoken!=null){
+                cred.setToken(mytoken.getToken());
             }
             //加入返回类数组
             creditsEntities.add(cred);
@@ -92,15 +105,18 @@ public class PersonControl {
      */
     @Transactional
     @PostMapping(value = "/persons", produces = "application/json")
-    public CreditsEntity person_add(@Validated @RequestBody CreditsEntity creditsEntity) {
+    public Object personAdd(@Validated @RequestBody CreditsEntity requestObj) {
+        if(verifyToken(requestObj.getToken())==false){
+            return "无请求权限，操作错误";
+        }
         //新增人员表
         Person person = new Person();
-        person.setName(creditsEntity.getPersonName());
-        person.setAge(creditsEntity.getPersonAge());
+        person.setName(requestObj.getPersonName());
+        person.setAge(requestObj.getPersonAge());
         person = personRepository.save(person);
 
         //新增人员-职位对照表
-        for (Position position : creditsEntity.getPosition()) {
+        for (Position position : requestObj.getPosition()) {
             int i = 0;
             Person_Position person_position = new Person_Position();
             person_position.setPerson(person);
@@ -108,20 +124,27 @@ public class PersonControl {
             ppRepository.save(person_position);
         }
         //返回人员信息
-        creditsEntity.setPersonId(person.getId());
-        return creditsEntity;
+        requestObj.setPersonId(person.getId());
+        return requestObj;
     }
 
 
     /*
     查询一个人员
      */
-    @GetMapping(value = "/persons/{id}")
-    public CreditsEntity find_person(@PathVariable("id") Integer id) {
+    @GetMapping(value = "/personsOne")
+    public Object findPerson(@RequestParam("id") Integer id,
+                             @RequestParam("access_token") String token) {
+        if(verifyToken(token)==false){
+            return "无请求权限，操作错误";
+        }
         //初始化返回信息类
         CreditsEntity creditsEntity = new CreditsEntity();
         //查询人员基本信息（不包括职业）
         Person person = personRepository.findOne(id);
+        if(person==null){
+            return "用户不存在";
+        }
         creditsEntity.setPersonId(id);
         creditsEntity.setPersonAge(person.getAge());
         creditsEntity.setPersonName(person.getName());
@@ -132,7 +155,10 @@ public class PersonControl {
             Position position = positionRepository.myFindPo(p.getPositionId());
             creditsEntity.getPosition().add(position);
         }
-
+        Token mytoken = tokenRepository.findOne(person.getId());
+        if(mytoken!=null){
+            creditsEntity.setToken(mytoken.getToken());
+        }
         return creditsEntity;
     }
 
@@ -141,8 +167,14 @@ public class PersonControl {
     更新一个人员
      */
     @PutMapping(value = "/persons", produces = "application/json")
-    public CreditsEntity person_update(@Validated @RequestBody CreditsEntity requestObj) {
+    public Object personUpdate(@Validated @RequestBody CreditsEntity requestObj) {
+        if(verifyToken(requestObj.getToken())==false){
+            return "无请求权限，操作错误";
+        }
         Person person = personRepository.findOne(requestObj.getPersonId());
+        if(person==null){
+            return "用户不存在";
+        }
         person.setName(requestObj.getPersonName());
         person.setAge(requestObj.getPersonAge());
 
@@ -164,15 +196,26 @@ public class PersonControl {
     /*
     删除一个人员
      */
-    @DeleteMapping(value = "/persons/{id}")
-    public void delete_person(@PathVariable("id") Integer id) {
+    @DeleteMapping(value = "/persons")
+    public String deletePerson(@RequestParam("id") Integer id,
+                               @RequestParam("access_token") String token) {
+        if(ppRepository.myFindPId(id)==null){
+            return "用户信息不存在，删除失败";
+        }
         List<Person_Position> person_positions = ppRepository.myFindPId(id);
         ppRepository.delete(person_positions);
+        tokenRepository.delete(id);
+        CreditsEntity creditsEntity = (CreditsEntity)map.get("user");
+        if(creditsEntity.getPersonId()==id){
+            PersonControl.map.put("token", "");
+            PersonControl.map.put("user", null);
+        }
+        return "删除成功";
     }
 
 
 
-    //数据库Token操作类
+    //数据库Token操作方法
     public CreditsEntity loginToken(UserLogin login){
         UUID uuid = UUID.randomUUID();
         Token token1 = new Token();
@@ -185,9 +228,11 @@ public class PersonControl {
                 tokenRepository.save(token1);
             }
         }
+        //保存Token
         token1.setId(login.getPerson().getId());
         token1.setToken(uuid.toString());
         tokenRepository.save(token1);
+
         //填充返回数据
         CreditsEntity result = new CreditsEntity();
         result.setPersonId(login.getPerson().getId());
@@ -198,12 +243,66 @@ public class PersonControl {
             result.getPosition().add(positionRepository.findOne(i.getPositionId()));
         }
         result.setToken(uuid.toString());
+
         //添加缓存数据
         map.put("token",token1.getToken());
         map.put("user",result);
+
+        //定时消除缓存
+        timeStart(token1.getId());
+
         return result;
     }
 
+    //定时清除缓存方法
+    public void timeStart(Integer id) {
+        TimerTask task = new TimerTask() {
+            @Override
+            public void run() {
+                // task to run goes here
+                tokenRepository.delete(id);
+                PersonControl.map.put("token", "");
+                PersonControl.map.put("user", null);
+                System.out.println("缓存清除成功");
+            }
+        };
+        Timer timer = new Timer();
+        timer.schedule(task, 300000);
+    }
 
+    //Token验证
+    public boolean verifyToken(String token){
+        CreditsEntity result = new CreditsEntity();
+        //缓存中寻找Token
+        if(map.get("token")!=null && map.get("token").equals(token)){
+            result = (CreditsEntity)map.get("user");
+            for(Position p:result.getPosition()){
+                if(p.getPositionName().equals("管理员")){
+                    return true;
+                }
+            }
+        }
+        //数据库中寻找Token
+        Token token1 = tokenRepository.findByToken(token);
+        if(token1==null){
+            return false;
+        }
+        List<Person_Position> list = ppRepository.myFindPId(token1.getId());
+        for(Person_Position i:list){
+            result.getPosition().add(positionRepository.findOne(i.getPositionId()));
+        }
+        for(Position p:result.getPosition()){
+            if(p.getPositionName().equals("管理员")){
+                return true;
+            }
+        }
+        return false;
+    }
 
+    @PostMapping(value = "/text", produces = "application/json")
+    public Boolean findMy(@Validated @RequestBody CreditsEntity requestObj){
+
+        return verifyToken(requestObj.getToken());
+
+    }
 }
